+++ scripts/process_restore_helpers.sh
@@ -1,20 +1,3 @@
-# default processes that are restored
-default_proc_list_option="@session-saver-default-processes"
-default_proc_list='vi vim emacs man less more tail top htop irssi irb pry "~rails console"'
-
-# User defined processes that are restored
-#  'false' - nothing is restored
-#  ':all:' - all processes are restored
-#
-# user defined list of programs that are restored:
-#  'my_program foo another_program'
-restore_processes_option="@session-saver-processes"
-restore_processes=""
-
-# Defines part of the user variable. Example usage:
-#   set -g @session-saver-strategy-vim "session"
-restore_process_strategy_option="@session-saver-strategy-"
-
 restore_pane_processes_enabled() {
 	local restore_processes="$(get_tmux_option "$restore_processes_option" "$restore_processes")"
 	if [ "$restore_processes" == "false" ]; then